+++ src/fs/fat/LongNameAssembly.hpp
@@ -8,7 +8,7 @@
 #include <optional>
 #include <vector>
 
-#include "revenant/fs/NameDecode.hpp"
+#include "revenant/core/Utf16Name.hpp"
 #include "revenant/fs/fat/DirectoryEntry.hpp"
 
 namespace revenant::fs::fat {